Add files via upload
The Category and Sub-category columns were updated to include the LEFT and RIGHT formulas.  These were added in yellow next to the existing ones from the prior submission.
The standard deviation and variance were updated to use population formulas instead of sample formulas.  I had used sample initially since I considered the successful and failed campaigns to be subsets of the full population of campaigns.  Updated values are in yellow.
Thanks.