Update README.md
Updates for file uploaded on 04Jan2023
If you are considering all the successful campaigns a POPULATION and all the failed campaigns a separate population, you can apply the formulas in yellow. I had previously used the sample formulas because I considered these SAMPLES from the full dataset.  See updated SD and variance output on the tab called "Stats for Number of Backers."

The LEFT and RIGHT formulas were shown in the new Parent Category and Sub-category columns added in the tab called "Crowdfunding"